Exclude twitter models from demo reset command.

+++ demo/management/commands/reset_demo.py
@@ -18,7 +18,7 @@ class Command(NoArgsCommand):
             a.startswith("mezzanine.")]
         for model in get_models():
             meta = model._meta
-            if meta.app_label in reset_apps:
+            if meta.app_label in reset_apps and meta.app_label != "twitter":
                 print "Flushing %s.%s" % (meta.app_label, meta.object_name)
                 model.objects.all().delete()
         call_command("loaddata", "initial_data", **options)
